==> ACUMAR/2_Procesamiento/1 - Generación de Capa EVP.ipynb <==
import pandas as pd
import geopandas as gpd

# Datos OSM

# importar capas
gdf_parks = gpd.read_file("../1_Datos/OSM/leisure_park_pol.geojson")
gdf_grass = gpd.read_file("../1_Datos/OSM/landuse_grass_pol.geojson")
gdf_grassland = gpd.read_file("../1_Datos/OSM/natural_grassland_pol.geojson")
gdf_forest = gpd.read_file("../1_Datos/OSM/landuse_forest_pol.geojson")
gdf_wood = gpd.read_file("../1_Datos/OSM/natural_wood_pol.geojson")
gdf_reserve = gpd.read_file("../1_Datos/OSM/leisure_naturereserve_pol.geojson")

# generación de campo categoría
gdf_parks['categoria'] = 'park'
gdf_grass['categoria'] = 'grass'
gdf_grassland['categoria'] = 'grassland'
gdf_forest['categoria'] = 'forest'
gdf_wood['categoria'] = 'wood'
gdf_reserve['categoria'] = 'naturereserve'

# concatenación de los gdfs
gdf_osm = pd.concat([gdf_parks, gdf_grass, gdf_grassland, gdf_forest, gdf_wood, gdf_reserve])
# generación del campo fuente y epv
gdf_osm['fuente'] = 1
gdf_osm['epv'] = 1
# filtro de las columnas de interés
gdf_osm = gdf_osm[['full_id', 'osm_id', 'leisure', 'epv', 'categoria', 'name', 'fuente', 'geometry']]

gdf_osm

# IGN

# importar capa
gdf_ign = gpd.read_file("../1_Datos/IGN/EspaciosVerdes.geojson")

# Filtro de las capas al área de interés

# importar capas de municipios y límite antrópico
gdf_muni = gpd.read_file("../1_Datos/ACUMAR/LIMITE_POLITICO.shp")
gdf_limite = gpd.read_file("../1_Datos/ACUMAR/LIMITE_ANTROPICO_POLIGONO.geojson")

# filtro de los municipios de interés 
gdf_muni = gdf_muni.loc[(gdf_muni.DISTRITO == "LOMAS DE ZAMORA") | (gdf_muni.DISTRITO == "LANUS")]

gdf_osm_muni = gpd.sjoin(gdf_osm, gdf_muni, how='left')
gdf_osm_muni = gdf_osm_muni.loc[gdf_osm_muni.index_right.notnull()]
gdf_osm_muni = gdf_osm_muni[['full_id', 'osm_id', 'leisure', 'epv', 'categoria', 'name', 'fuente', 'DISTRITO', 'geometry']]
print("Polígonos OSM únicos:", len(gdf_osm_muni.full_id.unique()) == len(gdf_osm_muni))
gdf_ign_muni = gpd.sjoin(gdf_ign, gdf_muni, how='left')
gdf_ign_muni = gdf_ign_muni.loc[gdf_ign_muni.index_right.notnull()]
gdf_ign_muni = gdf_ign_muni[['gid', 'entidad', 'objeto', 'fna', 'gna', 'nam', 'fdc', 'sag', 'DISTRITO', 'geometry']]
print("Polígonos IGN únicos:", len(gdf_ign_muni.gid.unique()) == len(gdf_ign_muni))

gdf_osm_lim = gpd.sjoin(gdf_osm_muni, gdf_limite, how='left')
gdf_osm_lim['limite'] = 1
gdf_osm_lim.loc[gdf_osm_lim.index_right.isnull(), 'limite'] = 0
gdf_osm_lim = gdf_osm_lim[['full_id', 'osm_id', 'leisure', 'epv', 'categoria', 'name', 'fuente', 'DISTRITO', 'limite', 'geometry']]
gdf_ign_lim = gpd.sjoin(gdf_ign_muni, gdf_limite, how='left')
gdf_ign_lim['limite'] = 1
gdf_ign_lim.loc[gdf_ign_lim.index_right.isnull(), 'limite'] = 0
gdf_ign_lim = gdf_ign_lim[['gid', 'entidad', 'objeto', 'fna', 'gna', 'nam', 'fdc', 'sag', 'DISTRITO', 'limite', 'geometry']]

gdf_osm_lim.to_file("./Capas_Output/OSM_epv.geojson", driver="GeoJSON")
gdf_ign_lim.to_file("./Capas_Output/IGM_epv.geojson", driver="GeoJSON")

# Registros OSM superpuestos

gdf_osm = gpd.read_file("./Capas_Output/OSM_epv.geojson")
gdf_ign = gpd.read_file("./Capas_Output/IGM_epv.geojson")

# Registro OSM superpuestos entre si
gdf_osm['superpuesto'] = 0

for i in range(len(gdf_osm)):
    gdf_registrosuperpuesto = gpd.sjoin(gdf_osm.loc[[i]], gdf_osm.loc[gdf_osm.index != i], how='left')

    if len(gdf_registrosuperpuesto.loc[gdf_registrosuperpuesto.full_id_right.notnull()]) > 0:
        gdf_osm.loc[i, 'superpuesto'] = 1

print("Polígonos sin superposición", len(gdf_osm.loc[gdf_osm['superpuesto']==0]))
print("Polígonos con superposición", len(gdf_osm.loc[gdf_osm['superpuesto']==1]))

# Cruce de capas

# Cruso capas y me quedo con las que no se superponen
gdf_ign_sjoin = gpd.sjoin(gdf_ign, gdf_osm, how='left')
print("Parcelas IGN:", len(gdf_ign_sjoin.loc[(gdf_ign_sjoin.index_right.isnull())]))
gdf_ign_sjoin = gdf_ign_sjoin[['gid', 'entidad', 'objeto', 'fna', 'gna', 'nam', 'fdc', 'sag',
       'DISTRITO_left', 'limite_left', 'geometry', 'index_right']]
gdf_ign_sjoin.rename(columns={'DISTRITO_left':'DISTRITO'}, inplace=True)
gdf_ign_sjoin.loc[(gdf_ign_sjoin.index_right.isnull())].to_file("./Capas_Output/IGN_sinintersect.geojson", driver="GeoJSON")

gdf_osm_sjoin = gpd.sjoin(gdf_osm, gdf_ign, how='left')
print("Parcelas OSM:", len(gdf_osm_sjoin.loc[(gdf_osm_sjoin.index_right.isnull())]))
gdf_osm_sjoin = gdf_osm_sjoin[['full_id', 'osm_id', 'leisure', 'epv', 'categoria', 'name', 'fuente',
       'DISTRITO_left', 'limite_left', 'geometry', 'superpuesto', 'index_right']]
gdf_osm_sjoin.rename(columns={'DISTRITO_left':'DISTRITO'}, inplace=True)
gdf_osm_sjoin.loc[(gdf_osm_sjoin.index_right.isnull())].to_file("./Capas_Output/OSM_sinintersect.geojson", driver="GeoJSON")

# cruce entre capas y seleccionar una intersección
gdf_ign['area_ign'] = gdf_ign.area
gdf_osm['area_osm'] = gdf_osm.area
gdf_ignosm_sjoin = gpd.sjoin(gdf_ign, gdf_osm, how='left')

""" gdf_ignosm_super = gdf_ignosm_sjoin.loc[(gdf_ignosm_sjoin.index_right.notnull())]
gdf_ignosm_super.reset_index(inplace=True, drop=True)
gdf_ign['super_b'] = 0
gdf_osm['super_b'] = 0

for i in range(len(gdf_ignosm_super)):
    if gdf_ignosm_super.loc[i, "area_ign"] > gdf_ignosm_super.loc[i, "area_osm"]:
        gdf_ign.loc[gdf_ign.gid == gdf_ignosm_super.loc[i, "gid"], 'super_b'] = 1
    else:
        gdf_osm.loc[gdf_osm.osm_id == gdf_ignosm_super.loc[i, "osm_id"], 'super_b'] = 1

gdf_ign.loc[gdf_ign.super_b == 1].to_file("./Capas_Output/IGN_intersect.geojson", driver="GeoJSON")
gdf_osm.loc[gdf_osm.super_b == 1].to_file("./Capas_Output/OSM_intersect.geojson", driver="GeoJSON")
print("Polígonos IGN:", len(gdf_ign.loc[gdf_ign.super_b == 1]))
print("Polígonos OSM:", len(gdf_osm.loc[gdf_osm.super_b == 1]))
print("Polígonos Totales:", len(gdf_ign.loc[gdf_ign.super_b == 1])+len(gdf_osm.loc[gdf_osm.super_b == 1])) """

gdf_ignosm_super = gdf_ignosm_sjoin.loc[(gdf_ignosm_sjoin.index_right.notnull())]
gdf_ignosm_super.reset_index(inplace=True, drop=True)
gdf_ign['super_b'] = 0
gdf_osm['super_b'] = 0

for i in range(len(gdf_ignosm_super)):
    if gdf_ignosm_super.loc[i, "area_osm"] > gdf_ignosm_super.loc[i, "area_ign"]:
        gdf_osm.loc[gdf_osm.osm_id == gdf_ignosm_super.loc[i, "osm_id"], 'super_b'] = 1
    else:
        gdf_ign.loc[gdf_ign.gid == gdf_ignosm_super.loc[i, "gid"], 'super_b'] = 1

gdf_ign.loc[gdf_ign.super_b == 1].to_file("./Capas_Output/IGN_intersect.geojson", driver="GeoJSON")
gdf_osm.loc[gdf_osm.super_b == 1].to_file("./Capas_Output/OSM_intersect.geojson", driver="GeoJSON")
print("Polígonos IGN:", len(gdf_ign.loc[gdf_ign.super_b == 1]))
print("Polígonos OSM:", len(gdf_osm.loc[gdf_osm.super_b == 1]))
print("Polígonos Totales:", len(gdf_ign.loc[gdf_ign.super_b == 1])+len(gdf_osm.loc[gdf_osm.super_b == 1]))

# Unión de capas

# Unión de capas
gdf_osm_inter = gpd.read_file("./Capas_Output/OSM_intersect.geojson")
gdf_osm_inter['fuente'] = 'OSM'
gdf_ign_inter = gpd.read_file("./Capas_Output/IGN_intersect.geojson")
gdf_ign_inter['fuente'] = 'IGM'
gdf_osm_sininter = gpd.read_file("./Capas_Output/OSM_sinintersect.geojson")
gdf_osm_sininter['fuente'] = 'OSM'
gdf_ign_sininter = gpd.read_file("./Capas_Output/IGN_sinintersect.geojson")
gdf_ign_sininter['fuente'] = 'IGM'
gdf_evp = pd.concat([gdf_osm_inter, gdf_ign_inter, gdf_osm_sininter, gdf_ign_sininter])

gdf_evp.loc[gdf_evp.full_id.isnull(), 'full_id'] = gdf_evp.loc[gdf_evp.full_id.isnull(), 'gid']
gdf_evp.loc[gdf_evp.fna.isnull(), 'fna'] = gdf_evp.loc[gdf_evp.fna.isnull(), 'name']
gdf_evp.loc[gdf_evp.categoria.isnull(), 'categoria'] = 'park'
gdf_evp.loc[gdf_evp.limite.isnull(), 'limite'] = gdf_evp.loc[gdf_evp.limite.isnull(), 'limite_left']

gdf_evp_clean = gdf_evp[['full_id', 'fuente', 'DISTRITO', 'categoria', 'gna', 'fna', 'superpuesto', 'limite', 'geometry']]
gdf_evp_clean.reset_index(inplace=True, drop=True)
gdf_evp_clean.rename(columns={'DISTRITO':'distrito', 'gna':'tipo', 'fna':'nombre'}, inplace=True)
gdf_evp_clean.head()

print(len(gdf_evp_clean))

# completar datos de nombre y tipo según IGN
gdf_esping_sjoin = gpd.sjoin(gdf_evp_clean, gdf_ign[['fna', 'gna', 'geometry']], how='left')
gdf_esping_sjoin.loc[gdf_esping_sjoin.fna.notnull(), 'nombre'] = gdf_esping_sjoin.loc[gdf_esping_sjoin.fna.notnull(), 'fna']
gdf_esping_sjoin.loc[gdf_esping_sjoin.gna.notnull(), 'tipo'] = gdf_esping_sjoin.loc[gdf_esping_sjoin.gna.notnull(), 'gna']
gdf_esping_sjoin = gdf_esping_sjoin[['full_id', 'fuente', 'distrito', 'categoria', 'tipo', 'nombre', 'superpuesto', 'limite', 'geometry']]
gdf_esping_sjoin

# Eliminar duplicados
gdf_esping_sjoin.drop_duplicates(subset='full_id', keep='first', inplace=True, ignore_index=True)

# Registro superpuestos entre si
gdf_esping_sjoin['superpuesto'] = 0

for i in range(len(gdf_esping_sjoin)):
    gdf_registrosuperpuesto = gpd.sjoin(gdf_esping_sjoin.loc[[i]], gdf_esping_sjoin.loc[gdf_esping_sjoin.index != i], how='left')

    if len(gdf_registrosuperpuesto.loc[gdf_registrosuperpuesto.full_id_right.notnull()]) > 0:
        gdf_esping_sjoin.loc[i, 'superpuesto'] = 1

print("Polígonos sin superposición", len(gdf_esping_sjoin.loc[gdf_esping_sjoin['superpuesto']==0]))
print("Polígonos con superposición", len(gdf_esping_sjoin.loc[gdf_esping_sjoin['superpuesto']==1]))

gdf_esping_sjoin.loc[gdf_esping_sjoin.tipo == 'Parque Municipal', 'tipo'] = 'Parque'

gdf_esping_sjoin['evp'] = 0
gdf_esping_sjoin.loc[gdf_esping_sjoin.categoria == 'park', 'evp'] = 1
gdf_esping_sjoin = gdf_esping_sjoin[['full_id', 'fuente', 'distrito', 'evp', 'categoria', 'tipo', 'nombre', 'superpuesto', 'limite', 'geometry']]
gdf_esping_sjoin

# Exportar capa
gdf_esping_sjoin.to_file("./Capas_Output/EspaciosPublicos.geojson", driver="GeoJSON")

# Cálculo de superficies

# Importar capa
gdf = gpd.read_file("./Capas_Output/EspaciosPublicos_Pau.geojson")
gdf = gdf.to_crs("EPSG:5347")
gdf['area_m2'] = gdf.area
gdf.area_m2 = round(gdf.area_m2, 2)
gdf = gdf.to_crs("EPSG:4326")
gdf.rename(columns={'Observaciones':'observaciones'}, inplace=True)
gdf = gdf[['full_id', 'fuente', 'distrito', 'evp', 'categoria', 'tipo', 'nombre','superpuesto', 'limite', 'area_m2', 'observaciones', 'geometry']]
gdf.head()

# corrección de superpuestos
# eliminar superpuestos que no deberían estar
print(len(gdf))
lista_e = ['w149700227', 'r5514857', 'w241322368', 'r2895800', '5902']
for e in lista_e:
    gdf.drop(gdf.loc[gdf.full_id == e].index, inplace=True)
gdf.reset_index(inplace=True, drop=True)
print(len(gdf), '-', len(lista_e))
# modificar valor del campo superpuesto para los polígonos que quedan
for s in ['1890', '12271', '6609', '5828', 'w167029793', 'w28184879', 'w224939240', 'w291374429', 'w302130495', 'w224939237', 'w224939239', 'w224939242']:
    gdf.loc[gdf.full_id == s, 'superpuesto'] = 0

# corregir polígono con error
gdf.loc[87, 'full_id'] = '10367'
gdf.loc[gdf.full_id == '10367', 'fuente'] = 'IGN'
gdf.loc[gdf.full_id == '10367', 'distrito'] = 'LANUS'
gdf.loc[gdf.full_id == '10367', 'categoria'] = 'park'
gdf.loc[gdf.full_id == '10367', 'limite'] = 0
gdf.loc[gdf.full_id == '10367']

# eliminar polígonos que no corresponden por ID
print(len(gdf))
lista_e2 = ['13221', '13562', 'w678895041']
for e in lista_e2:   # 'w678895041' => 10367
    gdf.drop(gdf.loc[gdf.full_id == e].index, inplace=True)
gdf.reset_index(inplace=True, drop=True)
print(len(gdf), '-', len(lista_e2))

# eliminar polígonos sin datos
print(len(gdf))
gdf = gdf.loc[gdf.fuente.notnull()] 
gdf.reset_index(inplace=True, drop=True)
print(len(gdf))

# Corrección de fuente
gdf.loc[gdf.fuente == 'IGM', 'fuente'] = 'IGN'

gdf.fuente.value_counts()

# exportar capa
gdf.to_file("./Capas_Output/EspaciosPublicos_v2.geojson", driver="GeoJSON")
