==> ACUMAR/2_Procesamiento/2 - Limpieza de capas editadas.ipynb <==
import pandas as pd
import geopandas as gpd

# Cargar Datos

# importar capas
gdf_OSMIGN = gpd.read_file("./CapasEditadas/EspaciosPublicos_v2.geojson")
gdf_UNSAM = gpd.read_file("./CapasEditadas/EspaciosPublicos_v2_UNSAM.2.geojson")

gdf_evp = pd.concat([gdf_OSMIGN, gdf_UNSAM])
gdf_evp.reset_index(inplace=True, drop=True)
print("Datos OSM+IGN:", len(gdf_OSMIGN))
print("Datos UNSAM:", len(gdf_UNSAM))
print("Datos Total:", len(gdf_OSMIGN)+len(gdf_UNSAM))
print("____________________________")
print("Datos Capa EPV:", len(gdf_evp))

# importar capas de municipios y límite antrópico
gdf_muni = gpd.read_file("../1_Datos/ACUMAR/LIMITE_POLITICO.shp")
gdf_limite = gpd.read_file("../1_Datos/ACUMAR/LIMITE_ANTROPICO_POLIGONO.geojson")

# importar barrios populares
gdf_barrios = gpd.read_file("../1_Datos/Renabap/BarriosPopulares_2022-05-15_info_publica.geojson")
gdf_barrios.head()

# Calcular superficie

gdf_evp = gdf_evp.to_crs("EPSG:5347")
gdf_evp['area_m2'] = gdf_evp.area
gdf_evp.area_m2 = round(gdf_evp.area_m2, 2)
gdf_evp = gdf_evp.to_crs("EPSG:4326")

# Limpiar el campo observaciones

# chequeo de fuentes
gdf_evp.fuente.unique()

# eliminar palabra nuevo de las observaciones de los datos de fuente UNSAM
gdf_evp.loc[gdf_evp.fuente == 'UNSAM', 'observaciones'] = gdf_evp.loc[gdf_evp.fuente == 'UNSAM', 'observaciones'].str.replace('nuevo. ', '')
gdf_evp.loc[gdf_evp.fuente == 'UNSAM', 'observaciones'] = gdf_evp.loc[gdf_evp.fuente == 'UNSAM', 'observaciones'].str.replace('nueva. ', '')
gdf_evp.loc[gdf_evp.fuente == 'UNSAM', 'observaciones'] = gdf_evp.loc[gdf_evp.fuente == 'UNSAM', 'observaciones'].str.replace('nuevo', '')
gdf_evp.loc[gdf_evp.fuente == 'UNSAM', 'observaciones'] = gdf_evp.loc[gdf_evp.fuente == 'UNSAM', 'observaciones'].str.replace('nueva', '')

# Crear nuevo campo bool y string para sistema

gdf_evp['sis_bool'] = 0
gdf_evp['sis_nam'] = None
gdf_evp = gdf_evp[['full_id', 'fuente', 'distrito', 'evp', 'categoria', 'tipo', 'nombre', 'sis_bool', 'sis_nam',
       'superpuesto', 'area_m2', 'limite', 'observaciones', 'geometry']]

# Crear nuevo campo bool y nombre de barrios popular

# presencia y nombre de barrio al que pertenece el espacio verde
gdf_evp_b = gpd.sjoin(gdf_evp, gdf_barrios, how='left')
gdf_evp = gdf_evp.merge(gdf_evp_b[['full_id', 'nombre_barrio']], on='full_id', how='left')
gdf_evp.rename(columns={'nombre_barrio':'barp_nam'}, inplace=True)
gdf_evp['barp_bool'] = 0
gdf_evp.loc[gdf_evp.barp_nam.notnull(), 'barp_bool'] = 1

# buffer de 100m y cruce de capas
gdf_evp_buffer = gdf_evp.copy()
gdf_evp_buffer = gdf_evp_buffer.to_crs("EPSG:5347")
gdf_evp_buffer.geometry = gdf_evp_buffer.buffer(100)
gdf_evp_buffer = gdf_evp_buffer.to_crs("EPSG:4326")

gdf_evp_buffer_b = gpd.sjoin(gdf_evp_buffer, gdf_barrios, how='left')

# cantidad y nombres de barrios a 100 de cada espacio verde
pd_bar100_size = pd.DataFrame(gdf_evp_buffer_b.loc[gdf_evp_buffer_b.nombre_barrio.notnull(), 'full_id'].value_counts())
pd_bar100_size.reset_index(inplace=True)
pd_bar100_size.rename(columns={'index':'full_id', 'full_id':'barp100_cant'}, inplace=True)
pd_bar100_nam = pd.DataFrame(gdf_evp_buffer_b.loc[gdf_evp_buffer_b.nombre_barrio.notnull()].groupby(['full_id'])['nombre_barrio'].apply(lambda x: (", ").join(list(set(x)))))
pd_bar100_nam.reset_index(inplace=True)

gdf_evp = gdf_evp.merge(pd_bar100_size[['full_id', 'barp100_cant']], on='full_id', how='left')
gdf_evp = gdf_evp.merge(pd_bar100_nam[['full_id', 'nombre_barrio']], on='full_id', how='left')
gdf_evp.rename(columns={'nombre_barrio':'barp100_nam'}, inplace=True)

# Generar valores de campo "full_id" para los datos UNSAM 

num = 1
for i in gdf_evp.loc[gdf_evp.fuente == "UNSAM"].index:
    num_str = str(num).zfill(5)
    gdf_evp.loc[i, 'full_id'] = "usm" + num_str
    num = num+1

# Limpieza de datos

# check polígonos eliminados
lista_e = ['w149700227', 'r5514857', 'w241322368', 'r2895800', '5902', '13221', '13562', 'w678895041', '107', 'w557932545', 'w557932547', 'w557932548',
        'r2467321', 'w160069211', 'w160069212', 'w160069213', 'w167029792', 'w160069208', 'w800490320', 'w160069206', 'w570792577', 'w510623913']

for e in lista_e:
    if len(gdf_evp.loc[gdf_evp.full_id == i]) > 0:
        print(i)

# check unique ids
len(gdf_evp) == len(gdf_evp.full_id.unique())

# check unique tipo
gdf_evp.tipo.unique()

# check unique distrito
gdf_evp.distrito.unique()

# recalcular distrito
gdf_evp_center = gdf_evp.copy()
gdf_evp_center.geometry = gdf_evp.centroid
gdf_evp_center_sjoin = gpd.sjoin(gdf_evp_center, gdf_muni[['DISTRITO', 'geometry']], how='left')
gdf_evp = gdf_evp.merge(gdf_evp_center_sjoin[['full_id', 'DISTRITO']], on='full_id', how='left')
gdf_evp['distrito'] = gdf_evp['DISTRITO']
gdf_evp.drop(columns='DISTRITO', inplace=True)

# cambiar el nombre de los distritos
gdf_evp.loc[gdf_evp.distrito == 'LANUS', 'distrito'] = 'Lanús'
gdf_evp.loc[gdf_evp.distrito != 'Lanús', 'distrito'] = 'Lomas de Zamora'
gdf_evp.distrito.unique()

# recalcular limite
gdf_evp = gpd.sjoin(gdf_evp, gdf_limite[['fid', 'geometry']], how='left')
gdf_evp['limite'] = 1
gdf_evp.loc[gdf_evp.index_right.isnull(), 'limite'] = 0

# ordenar columns y eliminar "superpuesto"
gdf_evp = gdf_evp[['full_id', 'fuente', 'distrito', 'evp', 'tipo', 'nombre', 'sis_bool', 'sis_nam',
        'area_m2', 'barp_nam', 'barp_bool', 'barp100_cant', 'barp100_nam', 'limite', 'observaciones', 
        'geometry']]

gdf_evp.to_file("./CapasEditadas/EspaciosPublicos_v4.geojson", driver="GeoJSON")

# Segundo Chequeo

gdf_evp = gpd.read_file("./CapasEditadas/EspaciosPublicos_v3-12-7.geojson")

# eliminar polígonos duplicados
print(len(gdf_evp))
gdf_evp = gdf_evp.loc[gdf_evp.full_id != 'usm00002']
gdf_evp = gdf_evp.loc[gdf_evp.full_id != 'usm00004']
gdf_evp = gdf_evp.loc[gdf_evp.full_id != 'usm00005']
gdf_evp.reset_index(inplace=True, drop=True)
print(len(gdf_evp))

# calcular áreas
gdf_evp = gdf_evp.to_crs("EPSG:5347")
gdf_evp['area_m2'] = gdf_evp.area
gdf_evp.area_m2 = round(gdf_evp.area_m2, 2)
gdf_evp = gdf_evp.to_crs("EPSG:4326")

# presencia y nombre de barrio al que pertenece el espacio verde
gdf_evp_b = gpd.sjoin(gdf_evp, gdf_barrios, how='left')
gdf_evp = gdf_evp.merge(gdf_evp_b[['full_id', 'nombre_barrio']], on='full_id', how='left')
gdf_evp.rename(columns={'nombre_barrio':'barp_nam'}, inplace=True)
gdf_evp['barp_bool'] = 0
gdf_evp.loc[gdf_evp.barp_nam.notnull(), 'barp_bool'] = 1

# buffer de 100m y cruce de capas
gdf_evp_buffer = gdf_evp.copy()
gdf_evp_buffer = gdf_evp_buffer.to_crs("EPSG:5347")
gdf_evp_buffer.geometry = gdf_evp_buffer.buffer(100)
gdf_evp_buffer = gdf_evp_buffer.to_crs("EPSG:4326")

gdf_evp_buffer_b = gpd.sjoin(gdf_evp_buffer, gdf_barrios, how='left')

# cantidad y nombres de barrios a 100 de cada espacio verde
pd_bar100_size = pd.DataFrame(gdf_evp_buffer_b.loc[gdf_evp_buffer_b.nombre_barrio.notnull(), 'full_id'].value_counts())
pd_bar100_size.reset_index(inplace=True)
pd_bar100_size.rename(columns={'index':'full_id', 'full_id':'barp100_cant'}, inplace=True)
pd_bar100_nam = pd.DataFrame(gdf_evp_buffer_b.loc[gdf_evp_buffer_b.nombre_barrio.notnull()].groupby(['full_id'])['nombre_barrio'].apply(lambda x: (", ").join(list(set(x)))))
pd_bar100_nam.reset_index(inplace=True)

gdf_evp = gdf_evp.merge(pd_bar100_size[['full_id', 'barp100_cant']], on='full_id', how='left')
gdf_evp = gdf_evp.merge(pd_bar100_nam[['full_id', 'nombre_barrio']], on='full_id', how='left')
gdf_evp.rename(columns={'nombre_barrio':'barp100_nam'}, inplace=True)
