==> ACUMAR/2_Procesamiento/2.1 - Limpieza de capas editadas.ipynb <==
import pandas as pd
import geopandas as gpd

# importar capas
gdf_evp = gpd.read_file("./CapasEditadas/EspaciosPublicos_v10.geojson")
gdf_evp.drop(columns=['sis_area_m2', 'barpop_nam', 'barpop_cant', 'bus50_cant', 'bus500_cant'], inplace=True)

# importar capas de municipios y límite antrópico
gdf_muni = gpd.read_file("../1_Datos/ACUMAR/LIMITE_POLITICO.shp")
gdf_limite = gpd.read_file("../1_Datos/ACUMAR/LIMITE_ANTROPICO_POLIGONO.geojson")

# importar capas de línea de colectivos
gdf_bus = gpd.read_file("./Capas_Output/LineasColectivos_v1.geojson")

# importar barrios populares
gdf_barrios = gpd.read_file("../1_Datos/Renabap/BarriosPopulares_2022-05-15_info_publica.geojson")
gdf_barrios.head()

# Crear nuevo campo bool y nombre de barrios popular

# presencia y nombre de barrio al que pertenece el espacio verde
""" gdf_evp_b = gpd.sjoin(gdf_evp, gdf_barrios, how='left')
gdf_evp_b.drop_duplicates(subset=['full_id'], inplace=True)
gdf_evp = gdf_evp.merge(gdf_evp_b[['full_id', 'nombre_barrio']], on='full_id', how='left')
gdf_evp.rename(columns={'nombre_barrio':'barp_nam'}, inplace=True)
gdf_evp['barp_bool'] = 0
gdf_evp.loc[gdf_evp.barp_nam.notnull(), 'barp_bool'] = 1 """

# buffer de 200m y cruce de capas
gdf_evp_buffer = gdf_evp.copy()
gdf_evp_buffer = gdf_evp_buffer.to_crs("EPSG:5347")
gdf_evp_buffer.geometry = gdf_evp_buffer.buffer(200)
gdf_evp_buffer = gdf_evp_buffer.to_crs("EPSG:4326")

gdf_evp_buffer_b = gpd.sjoin(gdf_evp_buffer, gdf_barrios, how='left')

# cantidad y nombres de barrios a 200 de cada espacio verde
pd_bar100_size = pd.DataFrame(gdf_evp_buffer_b.loc[gdf_evp_buffer_b.nombre_barrio.notnull(), 'full_id'].value_counts())
pd_bar100_size.reset_index(inplace=True)
pd_bar100_size.rename(columns={'index':'full_id', 'full_id':'barpop_cant'}, inplace=True)
pd_bar100_nam = pd.DataFrame(gdf_evp_buffer_b.loc[gdf_evp_buffer_b.nombre_barrio.notnull()].groupby(['full_id'])['nombre_barrio'].apply(lambda x: (", ").join(list(set(x)))))
pd_bar100_nam.reset_index(inplace=True)

gdf_evp = gdf_evp.merge(pd_bar100_size[['full_id', 'barpop_cant']], on='full_id', how='left')
gdf_evp = gdf_evp.merge(pd_bar100_nam[['full_id', 'nombre_barrio']], on='full_id', how='left')
gdf_evp.rename(columns={'nombre_barrio':'barpop_nam'}, inplace=True)
gdf_evp.barpop_cant.fillna(0, inplace=True)

# Generar valores de campo "full_id" para los datos UNSAM 

""" num = 1
for i in gdf_evp.loc[gdf_evp.fuente == "UNSAM"].index:
    num_str = str(num).zfill(5)
    gdf_evp.loc[i, 'full_id'] = "usm" + num_str
    num = num+1 """

# Suma de área de sistemas líneales

gdf_evp.loc[gdf_evp.sis_nam == '', 'sis_nam'] = None

df_sissup = pd.DataFrame(gdf_evp.loc[(gdf_evp.sis_nam.notnull())&(gdf_evp.sis_nam != '')].groupby("sis_nam")['area_m2'].sum()).reset_index()
df_sissup.rename(columns={'area_m2':'sis_area_m2'}, inplace=True)

print(len(gdf_evp))
gdf_evp = gdf_evp.merge(df_sissup, on='sis_nam', how='left')
print(len(gdf_evp))

# Cruce con líneas de colectivos

# buffer de 50m y cruce capa capa de líenas de colectivos
gdf_evp_buffer_50m = gdf_evp.copy()
gdf_evp_buffer_50m = gdf_evp_buffer_50m.to_crs("EPSG:5347")
gdf_evp_buffer_50m.geometry = gdf_evp_buffer_50m.buffer(50)
gdf_evp_buffer_50m = gdf_evp_buffer_50m.to_crs("EPSG:4326")

gdf_evp_buffer_50m_sj = gpd.sjoin(gdf_evp_buffer_50m, gdf_bus, how='left')
df_bus50 = pd.DataFrame(gdf_evp_buffer_50m_sj.loc[gdf_evp_buffer_50m_sj.full_id_right.notnull(), 'full_id_left'].value_counts()).reset_index().rename(columns={'index':'full_id', 'full_id_left':'bus50_cant'})

# buffer de 500m y cruce capa capa de líenas de colectivos
gdf_evp_buffer_500m = gdf_evp.copy()
gdf_evp_buffer_500m = gdf_evp_buffer_500m.to_crs("EPSG:5347")
gdf_evp_buffer_500m.geometry = gdf_evp_buffer_500m.buffer(500)
gdf_evp_buffer_500m = gdf_evp_buffer_500m.to_crs("EPSG:4326")

gdf_evp_buffer_500m_sj = gpd.sjoin(gdf_evp_buffer_500m, gdf_bus, how='left')
df_bus500 = pd.DataFrame(gdf_evp_buffer_500m_sj.loc[gdf_evp_buffer_500m_sj.full_id_right.notnull(), 'full_id_left'].value_counts()).reset_index().rename(columns={'index':'full_id', 'full_id_left':'bus500_cant'})

print(len(gdf_evp))
gdf_evp = gdf_evp.merge(df_bus50, on='full_id', how='left')
gdf_evp = gdf_evp.merge(df_bus500, on='full_id', how='left')
gdf_evp.bus50_cant.fillna(0, inplace=True)
gdf_evp.bus500_cant.fillna(0, inplace=True)
print(len(gdf_evp))

# Limpieza de datos

# check unique ids
len(gdf_evp) == len(gdf_evp.full_id.unique())

# check unique tipo
gdf_evp.tipo.unique()

print(gdf_evp.loc[gdf_evp.tipo == '-', 'evp'].unique())
print(gdf_evp.loc[gdf_evp.tipo == 'Cancha', 'evp'].unique())
print(gdf_evp.loc[gdf_evp.tipo == 'Separador vial', 'evp'].unique())
print(gdf_evp.loc[gdf_evp.evp == 0, 'tipo'].unique())

# check unique distrito
gdf_evp.distrito.unique()

gdf_evp[['full_id', 'fuente', 'distrito', 'evp', 'tipo', 'nombre', 'limite', 'area_m2', 
       'sis_bool', 'sis_nam', 'sis_area_m2', 'barpop_cant', 'barpop_nam', 'bus50_cant', 'bus500_cant', 
       'observaciones', 'geometry']].to_file("./CapasEditadas/EspaciosPublicos_v10.1.geojson", driver="GeoJSON")

# Capas para mapas

gdf_evp[['full_id', 'fuente', 'distrito', 'evp', 'tipo', 'nombre', 'limite', 'area_m2', 
       'sis_bool', 'sis_nam', 'sis_area_m2', 'barpop_cant', 'barpop_nam', 'bus50_cant', 'bus500_cant', 
       'observaciones', 'geometry']].to_file("./Mapas/MapLayouts/shp/EspaciosPublicos.shp")

gdf_limite = gdf_limite.dissolve('Layer')
gdf_limite = gdf_limite[['geometry']] 
gdf_limite.to_file("./Mapas/MapLayouts/shp/LimiteAntropico.shp")

gdf_muni_inersect = gdf_muni.overlay(gdf_limite, how='intersection')
gdf_muni_inersect.to_file("./Mapas/MapLayouts/shp/LimitePolíticoAntropico.shp")

gdf_muni.to_file("./Mapas/MapLayouts/shp/LimitePolítico.shp")
