==> ACUMAR/2_Procesamiento/3 - Datos de EVPs.ipynb <==
import pandas as pd
import geopandas as gpd

gdf_evp = gpd.read_file("./CapasEditadas/EspaciosPublicos_v10.1.geojson", driver="GeoJSON")

print("cantidad total de epv:", len(gdf_evp))
print("-- Partido: --")
for partido in list(gdf_evp.distrito.unique()):
    print(" ", "cantidad de epv en "+partido+":", len(gdf_evp.loc[gdf_evp.distrito == partido]))
print("-- Límite: --")
print(" ", "cantidad de epv dentro del límite antrópico:", len(gdf_evp.loc[gdf_evp.limite == 1]))
print(" ", "cantidad de epv fuera del límite antrópico:", len(gdf_evp.loc[gdf_evp.limite == 0]))
print("-- Espacios Lineales: --")
print(" ", "cantidad de epv en sistema lineal:", len(gdf_evp.loc[gdf_evp.sis_bool == 1]))
print(" ", "cantidad sistemas linelaes:", len(gdf_evp.loc[gdf_evp.sis_bool == 1, 'sis_nam'].unique()))
print("-- Tipología: --")
for tipo in list(gdf_evp.tipo.unique()):
    print(" ", tipo+":", len(gdf_evp.loc[gdf_evp.tipo == tipo]))
print("-- Fuente: --")
for source in list(gdf_evp.fuente.unique()):
    print(" ", source+":", len(gdf_evp.loc[gdf_evp.fuente == source]))
