==> ACUMAR/2_Procesamiento/4 - Zonas para encuestas.ipynb <==
import pandas as pd
import geopandas as gpd

# importar capas de ríos
gdf_riachuelo = gpd.read_file("../1_Datos/ACUMAR/RED_HIDRICA_CMR.shp")
gdf_limpolitico = gpd.read_file("../1_Datos/ACUMAR/LIMITE_POLITICO.shp")
gdf_zonas = gpd.read_file("../1_Datos/UNSAM/Zonificación_Localidades_L y LZ.geojson")

# Zona de riachuelo

gdf_riachuelo = gdf_riachuelo.loc[gdf_riachuelo.OBSERVACIO == 'RECTIFICACION']
gdf_riachuelo_buffer = gdf_riachuelo.to_crs("EPSG:5347")
gdf_riachuelo_buffer.geometry = gdf_riachuelo_buffer.buffer(800,   )
gdf_riachuelo_buffer = gdf_riachuelo_buffer.to_crs("EPSG:4326")
gdf_riachuelo_buffer = gdf_riachuelo_buffer.dissolve('OBSERVACIO')

# intersect by limpolitico
gdf_limpolitico = gdf_limpolitico.loc[(gdf_limpolitico.DISTRITO == 'LANUS') | (gdf_limpolitico.DISTRITO == 'LOMAS DE ZAMORA')]
gdf_zona_riachuelo = gdf_riachuelo_buffer.overlay(gdf_limpolitico, how='intersection')

gdf_zona_riachuelo.rename(columns={'DISTRITO':'municipio'}, inplace = True)
gdf_zona_riachuelo['id'] = 15
gdf_zona_riachuelo['localidad'] = 'x'
gdf_zona_riachuelo.loc[gdf_zona_riachuelo.municipio == 'LANUS', 'localidad'] = 'SIRGA - Lanús'
gdf_zona_riachuelo.loc[gdf_zona_riachuelo.municipio == 'LANUS', 'municipio'] = 'Lanús'
gdf_zona_riachuelo.loc[gdf_zona_riachuelo.municipio == 'LOMAS DE ZAMORA', 'id'] = 16
gdf_zona_riachuelo.loc[gdf_zona_riachuelo.municipio == 'LOMAS DE ZAMORA', 'localidad'] = 'SIRGA - Lomas de Zamora'
gdf_zona_riachuelo.loc[gdf_zona_riachuelo.municipio == 'LOMAS DE ZAMORA', 'municipio'] = 'Lomas de Zamora'
gdf_zona_riachuelo = gdf_zona_riachuelo[['id', 'municipio', 'localidad', 'geometry']]

# Merge all zonas

gdf_zonas_dif = gdf_zonas.overlay(gdf_riachuelo_buffer, how='difference')

gdf_zonas_complete = pd.concat([gdf_zonas_dif, gdf_zona_riachuelo])
gdf_zonas_complete.reset_index(inplace=True, drop=True)
gdf_zonas_complete

gdf_zonas_complete.to_file("./CapasEditadas/Zonificación_Localidades_LyLZ_SIRGA.geojson", driver="GeoJSON")
