==> ACUMAR/2_Procesamiento/5 - Buffer de EP.ipynb <==
# Ep buffer
### This notebook create a buffer for each EP type, acording to its radius of influence.

# import libraries
import pandas as pd
import geopandas as gpd

## Load file

df = gpd.read_file("./CapasEditadas/EspaciosPublicos_v11.geojson")
df.head(3)

## Buffers according to radius

# unique EP types
df.tipo.unique()

# EP radius
radius = {
    'Plazoleta': 400,
    'Plaza': 600,
    'Parque': 800,
    'Parque con equipamiento': 2000
}

df_plazoleta = df.loc[df.tipo == 'Plazoleta']
df_plazoleta.reset_index(inplace=True, drop=True)
df_plazoleta = df_plazoleta.to_crs("EPSG:5347")
df_plazoleta.geometry = df_plazoleta.buffer(radius['Plazoleta'])
df_plazoleta = df_plazoleta.dissolve(by='tipo')
df_plazoleta = df_plazoleta[['geometry']]
df_plazoleta.to_file("./BufferEP/EP_Plazoletas_buffer"+str(radius['Plazoleta'])+".geojson", driver='GeoJSON')

df_plaza = df.loc[df.tipo == 'Plaza']
df_plaza.reset_index(inplace=True, drop=True)
df_plaza = df_plaza.to_crs("EPSG:5347")
df_plaza.geometry = df_plaza.buffer(radius['Plaza'])
df_plaza = df_plaza.dissolve(by='tipo')
df_plaza = df_plaza[['geometry']]
df_plaza.to_file("./BufferEP/EP_Plaza_buffer"+str(radius['Plaza'])+".geojson", driver='GeoJSON')

df_parque = df.loc[df.tipo == 'Parque']
df_parque.reset_index(inplace=True, drop=True)
df_parque = df_parque.to_crs("EPSG:5347")
df_parque.geometry = df_parque.buffer(radius['Parque'])
df_parque = df_parque.dissolve(by='tipo')
df_parque = df_parque[['geometry']]
df_parque.to_file("./BufferEP/EP_Parque_buffer"+str(radius['Parque'])+".geojson", driver='GeoJSON')

df_parqueconequip = df.loc[df.tipo == 'Parque con equipamiento']
df_parqueconequip.reset_index(inplace=True, drop=True)
df_parqueconequip = df_parqueconequip.to_crs("EPSG:5347")
df_parqueconequip.geometry = df_parqueconequip.buffer(radius['Parque con equipamiento'])
df_parqueconequip = df_parqueconequip.dissolve(by='tipo')
df_parqueconequip = df_parqueconequip[['geometry']]
df_parqueconequip.to_file("./BufferEP/EP_ParqueConEquip_buffer"+str(radius['Parque con equipamiento'])+".geojson", driver='GeoJSON')
