==> ACUMAR/2_Procesamiento/7 - Datos por partido y localidad.ipynb <==
# import libraries
import pandas as pd
import geopandas as gpd

## Load file

df = gpd.read_file("./CapasEditadas/EspaciosPublicos_v11.geojson")
df['evp'] = 0

# EP radius
df.loc[df.tipo == 'Plazoleta', 'evp'] = 1
df.loc[df.tipo == 'Plaza', 'evp'] = 1
df.loc[df.tipo == 'Parque', 'evp'] = 1
df.loc[df.tipo == 'Parque con equipamiento', 'evp'] = 1

df_ep = df.loc[(df.limite == 1)&(df.evp == 1)].reset_index(drop=True)
df_ep.head(3)

# importar capas de municipios y límite antrópico
gdf_muni = gpd.read_file("../1_Datos/ACUMAR/LIMITE_POLITICO.shp")
gdf_antro = gpd.read_file("../1_Datos/ACUMAR/LIMITE_ANTROPICO_POLIGONO.geojson")
# filtro de los municipios de interés 
gdf_muni = gdf_muni.loc[(gdf_muni.DISTRITO == "LOMAS DE ZAMORA") | (gdf_muni.DISTRITO == "LANUS")]

# importar capas de localidades
gdf_local = gpd.read_file("../1_Datos/Zonificación_Localidades_L y LZ.geojson")

# number and area by distrito

df_ep.groupby('distrito')['area_m2'].agg(['size', 'sum']).reset_index()

# number and area by localidad

# clip localidades
df_dif = gpd.overlay(gdf_local, gdf_antro[['geometry']], how='intersection')
#df_dif_sj = gpd.overlay(df_dif, df_ep[['full_id', 'area_m2', 'geometry']], how='intersection')
#df_dif_sj.area_m2 = df_dif_sj.to_crs("EPSG:5347").area
df_dif_sj = gpd.sjoin(df_dif, df_ep[['full_id', 'area_m2', 'geometry']])
df_dif_sj.drop_duplicates(subset="full_id", inplace=True)

df_dif_data = df_dif_sj.groupby('localidad')['area_m2'].agg(['size', 'sum']).reset_index()
gdf_local_data = pd.merge(df_dif, df_dif_data, on='localidad', how='left')
gdf_local_data.to_file("./Indice_m2/EP_localidad_antro.geojson", driver='GeoJSON')
gdf_local_data[['municipio', 'localidad', 'size', 'sum']].to_csv("./Indice_m2/EP_localidad_antro.csv")

gdf_local_data[['municipio', 'localidad', 'size', 'sum']]

# Datos

print("area", df.loc[(df.limite == 1), 'area_m2'].sum()/1000000, 'Mm²')
print("% area", 100*df.loc[(df.limite == 1), 'area_m2'].sum()/gdf_local_data.to_crs("EPSG:5347").area.sum(), '%')
print("area", df_ep.area_m2.sum()/1000000, 'Mm²')
print("% area", 100*df_ep.area_m2.sum()/df.loc[(df.limite == 1), 'area_m2'].sum(), '%')

sup_data = df_ep.groupby(['tipo'])['area_m2'].agg(['size', 'sum'])
sup_data['porc'] = 100* sup_data['sum'] / df_ep.area_m2.sum()
sup_data
